deprecated_ingestion_api/requirements.txt
boto3==1.42.54
Flask==3.1.3
PyYAML==6.0.3